+++ CHANGES.md
@@ -6,6 +6,7 @@
 * Update fix for API response header for JSON content type and the return of JSONP data to updated package (ref:hacks.txt)
 * Update Requests library 2.6.2 to 2.7.0 (ab1f493)
 * Update change to suppress HTTPS verification InsecureRequestWarning to updated package (ref:hacks.txt)
+* Change to consolidate cache database migration code
 
 
 ### 0.9.0 (2015-05-18 14:33:00 UTC)